+++ src/components/Hero/Hero.test.tsx
@@ -19,7 +19,7 @@ export const otherSubHeaderText =
   "Visit us today.";
 const imageProps = {
   alt: "Image example",
-  src: "//placekitten.com/800/400",
+  src: "//loremflickr.com/800/400/new+york+public+library",
 };
 
 describe("Hero accessbility tests", () => {
@@ -29,7 +29,7 @@ describe("Hero accessbility tests", () => {
         heroType="primary"
         heading={<Heading level="h1" id="a11y-hero" text="Hero Primary" />}
         subHeaderText="Example Subtitle"
-        backgroundImageSrc="//placekitten.com/1600/800"
+        backgroundImageSrc="//loremflickr.com/1600/800/new+york+public+library"
       />
     );
     expect(await axe(container)).toHaveNoViolations();
@@ -61,7 +61,7 @@ describe("Hero accessbility tests", () => {
   it("passes for type Campaign", async () => {
     const { container } = render(
       <Hero
-        backgroundImageSrc="//placekitten.com/g/2400/800"
+        backgroundImageSrc="//loremflickr.com/g/2400/800/new+york+public+library"
         heroType="campaign"
         heading={<Heading level="h1" id="a11y-hero" text="Hero Campaign" />}
         imageProps={imageProps}
@@ -90,7 +90,7 @@ describe("Hero", () => {
         heroType="primary"
         heading={<Heading level="h1" id="primary-hero" text="Hero Primary" />}
         subHeaderText="Example Subtitle"
-        backgroundImageSrc="//placekitten.com/1600/800"
+        backgroundImageSrc="//loremflickr.com/1600/800/new+york+public+library"
       />
     );
 
@@ -98,7 +98,7 @@ describe("Hero", () => {
     expect(screen.getByText("Example Subtitle")).toBeInTheDocument();
     expect(screen.getByTestId("hero")).toHaveAttribute(
       "style",
-      "background-image: url(//placekitten.com/1600/800);"
+      "background-image: url(//loremflickr.com/1600/800/new+york+public+library);"
     );
   });
 
@@ -119,7 +119,7 @@ describe("Hero", () => {
     expect(screen.getByRole("img")).toBeInTheDocument();
     expect(screen.getByRole("img")).toHaveAttribute(
       "src",
-      "//placekitten.com/800/400"
+      "//loremflickr.com/800/400/new+york+public+library"
     );
   });
 
@@ -140,7 +140,7 @@ describe("Hero", () => {
   it("renders Campaign Hero", () => {
     render(
       <Hero
-        backgroundImageSrc="//placekitten.com/1600/800"
+        backgroundImageSrc="//loremflickr.com/1600/800/new+york+public+library"
         heroType="campaign"
         heading={<Heading level="h1" id="campaign-hero" text="Hero Campaign" />}
         imageProps={imageProps}
@@ -153,12 +153,12 @@ describe("Hero", () => {
     // @TODO: This test needs to be rewritten to target a pseudo element.
     // expect(screen.getByTestId("hero")).toHaveAttribute(
     //   "style",
-    //   "background-image: url(//placekitten.com/1600/800);"
+    //   "background-image: url(//loremflickr.com/1600/800/new+york+public+library);"
     // );
     expect(screen.getByRole("img")).toBeInTheDocument();
     expect(screen.getByRole("img")).toHaveAttribute(
       "src",
-      "//placekitten.com/800/400"
+      "//loremflickr.com/800/400/new+york+public+library"
     );
   });
 
@@ -175,7 +175,7 @@ describe("Hero", () => {
     expect(screen.getByRole("img")).toBeInTheDocument();
     expect(screen.getByRole("img")).toHaveAttribute(
       "src",
-      "//placekitten.com/800/400"
+      "//loremflickr.com/800/400/new+york+public+library"
     );
   });
 
@@ -190,7 +190,7 @@ describe("Hero", () => {
             text="Hero with Custom Colors"
           />
         }
-        backgroundImageSrc="//placekitten.com/1600/800"
+        backgroundImageSrc="//loremflickr.com/1600/800/new+york+public+library"
         foregroundColor="#123456"
         backgroundColor="#654321"
       />
@@ -239,7 +239,7 @@ describe("Hero", () => {
     const warn = jest.spyOn(console, "warn");
     render(
       <Hero
-        backgroundImageSrc="//placekitten.com/1600/800"
+        backgroundImageSrc="//loremflickr.com/1600/800/new+york+public+library"
         heroType="primary"
         imageProps={{ src: imageProps.src }}
       />
@@ -263,7 +263,7 @@ describe("Hero", () => {
 
     rerender(
       <Hero
-        backgroundImageSrc="//placekitten.com/1600/800"
+        backgroundImageSrc="//loremflickr.com/1600/800/new+york+public+library"
         heroType="primary"
         heading={heading}
         imageProps={imageProps}
@@ -296,7 +296,7 @@ describe("Hero", () => {
 
     rerender(
       <Hero
-        backgroundImageSrc="//placekitten.com/1600/800"
+        backgroundImageSrc="//loremflickr.com/1600/800/new+york+public+library"
         heroType="secondary"
         heading={heading}
         imageProps={imageProps}
@@ -358,7 +358,7 @@ describe("Hero", () => {
 
     rerender(
       <Hero
-        backgroundImageSrc="//placekitten.com/1600/800"
+        backgroundImageSrc="//loremflickr.com/1600/800/new+york+public+library"
         heroType="tertiary"
         heading={heading}
         subHeaderText={otherSubHeaderText}
@@ -377,7 +377,7 @@ describe("Hero", () => {
     );
     const { rerender } = render(
       <Hero
-        backgroundImageSrc="//placekitten.com/g/2400/800"
+        backgroundImageSrc="//loremflickr.com/g/2400/800/new+york+public+library"
         heroType="campaign"
         heading={heading}
         imageProps={imageProps}
@@ -410,7 +410,7 @@ describe("Hero", () => {
         heroType="campaign"
         heading={heading}
         subHeaderText={otherSubHeaderText}
-        backgroundImageSrc="//placekitten.com/g/2400/800"
+        backgroundImageSrc="//loremflickr.com/g/2400/800/new+york+public+library"
         locationDetails={<>Some location details.</>}
       />
     );
@@ -472,7 +472,7 @@ describe("Hero", () => {
 
     rerender(
       <Hero
-        backgroundImageSrc="//placekitten.com/g/2400/800"
+        backgroundImageSrc="//loremflickr.com/g/2400/800/new+york+public+library"
         heroType="fiftyFifty"
         imageProps={imageProps}
         subHeaderText={otherSubHeaderText}
@@ -491,7 +491,7 @@ describe("Hero", () => {
           heroType="primary"
           heading={<Heading level="h1" id="primary-hero" text="Hero Primary" />}
           subHeaderText="Example Subtitle"
-          backgroundImageSrc="//placekitten.com/1600/800"
+          backgroundImageSrc="//loremflickr.com/1600/800/new+york+public+library"
         />
       )
       .toJSON();
@@ -570,7 +570,7 @@ describe("Hero", () => {
     const campaign = renderer
       .create(
         <Hero
-          backgroundImageSrc="//placekitten.com/g/2400/800"
+          backgroundImageSrc="//loremflickr.com/g/2400/800/new+york+public+library"
           heroType="campaign"
           heading={
             <Heading level="h1" id="campaign-hero" text="Hero Campaign" />
@@ -606,7 +606,7 @@ describe("Hero", () => {
           heroType="primary"
           heading={<Heading level="h1" id="chakra" text="Hero Primary" />}
           subHeaderText="Example Subtitle"
-          backgroundImageSrc="//placekitten.com/1600/800"
+          backgroundImageSrc="//loremflickr.com/1600/800/new+york+public+library"
           p="20px"
           color="ui.error.primary"
         />
@@ -618,7 +618,7 @@ describe("Hero", () => {
           heroType="primary"
           heading={<Heading level="h1" id="props" text="Hero Primary" />}
           subHeaderText="Example Subtitle"
-          backgroundImageSrc="//placekitten.com/1600/800"
+          backgroundImageSrc="//loremflickr.com/1600/800/new+york+public+library"
           data-testid="props"
         />
       )
@@ -641,7 +641,7 @@ describe("Hero", () => {
     const ref = React.createRef<HTMLDivElement>();
     const { container } = render(
       <Hero
-        backgroundImageSrc="//placekitten.com/1600/800"
+        backgroundImageSrc="//loremflickr.com/1600/800/new+york+public+library"
         heroType="primary"
         heading={<Heading level="h1" id="primary-hero" text="Hero Primary" />}
         ref={ref}