+++ src/components/Image/Image.test.tsx
@@ -44,7 +44,7 @@ describe("Image", () => {
 
   // @TODO - test when it does come into view.
   it("does not render an image src when `isLazy` is true until it is 'inView'", () => {
-    const src = "//placekitten.com/500/200";
+    const src = "//loremflickr.com/500/200/new+york+public+library";
     const { container } = render(<Image alt="" isLazy src={src} />);
 
     // Mock that the image is not in view through the IntersectionObserver object.