+++ src/components/Menu/Menu.stories.tsx
@@ -100,7 +100,11 @@ const imageListItems: ListItemsData[] = [
     type: "action",
     id: "item-title-1",
     label: "I'm item 1",
-    media: { type: "image", src: "//placekitten.com/300/300", alt: "" },
+    media: {
+      type: "image",
+      src: "//loremflickr.com/300/300/new+york+public+library",
+      alt: "",
+    },
     onClick: () => {
       console.log("Item Title 1 clicked");
     },
@@ -109,7 +113,11 @@ const imageListItems: ListItemsData[] = [
     type: "action",
     id: "item-title-2",
     label: "I'm item 2",
-    media: { type: "image", src: "//placekitten.com/300/300", alt: "" },
+    media: {
+      type: "image",
+      src: "//loremflickr.com/300/300/new+york+public+library",
+      alt: "",
+    },
     onClick: () => {
       console.log("Item Title 2 clicked");
     },
@@ -118,7 +126,11 @@ const imageListItems: ListItemsData[] = [
     type: "action",
     id: "item-title-3",
     label: "I'm item 3",
-    media: { type: "image", src: "//placekitten.com/300/300", alt: "" },
+    media: {
+      type: "image",
+      src: "//loremflickr.com/300/300/new+york+public+library",
+      alt: "",
+    },
     onClick: () => {
       console.log("Item Title 3 clicked");
     },