+++ src/components/Template/Template.stories.tsx
@@ -269,13 +269,14 @@ export const FullExampleWithTemplateChildrenComponents: Story = {
               ]}
             />
             <Hero
-              backgroundImageSrc="//placekitten.com/g/2400/800"
+              backgroundImageSrc="//loremflickr.com/g/2400/800/new+york+public+library"
               heroType="campaign"
               heading={<Heading level="h1" id="1" text="Hero Campaign" />}
               imageProps={{
                 alt: "Image example",
-                src: "//placekitten.com/800/400",
+                src: "//loremflickr.com/800/400/new+york+public+library",
               }}
+              isDarkBackgroundImage
               subHeaderText={otherSubHeaderText}
             />
           </TemplateHeader>
@@ -345,16 +346,13 @@ export const FullExampleWithTemplateChildrenComponents: Story = {
               imageProps={{
                 alt: "Alt text",
                 aspectRatio: "square",
-                size: "small",
-                src: "//placekitten.com/400/200",
+                size: "default",
+                src: "//loremflickr.com/400/200/new+york+public+library",
               }}
               isCentered
             >
-              <CardHeading size="heading4" id="heading1">
-                Small Animal Image
-              </CardHeading>
-              <CardHeading level="h3" size="heading6" id="heading2">
-                Animal info
+              <CardHeading size="heading4" subtitle="Animal info" id="heading1">
+                Library Image
               </CardHeading>
               <CardContent>
                 Vestibulum id ligula porta felis euismod semper. Nulla vitae