+++ CHANGELOG.md
@@ -8,6 +8,10 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 
 ## Prerelease
 
+### Updates
+
+- Updates placeholder images to come from `loremflickr.com`, rather than `placekitten.com`.
+
 ## 3.1.0 (April 11, 2024)
 
 ### Adds